Levels
This commit adds levels, that is:
1) Ability to specify number and type of enemies per level
2) A button to move to next level after end of each level

This required a refactor to how interaction works.
The way interaction was done before was causing problems.
If everything inherits from InteractableItem, it assumes it has
physical properties. Not everything that's interactable is physical,
eg. a button is just a kinematic object floating in space.

Since WandController can't collect objects in the scene based on
whether or not it implements an interface (GetComponent()'s limitation)
everything interacts from InteractableBase which handles inputs
as a virtual function with NOOP implementation.

Then, physical objects can subclass an abstract class that extends
InteractableBase, which has the physical properties from InteractableItem,
while LevelStartButton can override trigger button to actually do
something.

This commit also adds number of tweaks, most which won't work
from pulling from the git repo because of unity assets that aren't being
included. At this point, the enemies are more than just cubes, and t pare animated.